solution/requirements.txt
beautifulsoup4==4.6.1
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
idna==2.7
python-dateutil==2.7.3
requests==2.19.1
six==1.11.0
urllib3==1.23
